+++ src/main/java/duke/Duke.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Main entry point of the program.
  */
