+++ src/main/java/duke/Parser.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.exceptions.DukeException;
+import duke.exceptions.DukeUnknownCommandException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
